refactor: extract general chat restrictions
